backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
pydantic-settings==2.2.1
python3-saml==1.16.0
boto3==1.34.34
psutil==5.9.8
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
asyncssh==2.14.2
aiofiles==23.2.1
structlog==24.1.0
pytz==2024.1
